fix: restore Windows daemon support (PR #163 regression + new fixes)
The v0.3 refactor to CompozyOS lost the Windows platform-specific files
merged in PR #163 (fix: add Windows daemon support). Reproduces the
original issues on Windows 11 with the v0.3.0-beta.2 release binary:

Original PR #163 regressions re-fixed:
- #148 daemon start fails: "another process has locked a portion of the
  file" (LockFileEx byte-range lock prevents writeLockPID)
- #170 same, root-caused as LockFileEx conflict with in-process write
- #157 daemon start fails: "invalid uri authority: C:" (sqliteDSN passes
  drive-letter path to url.URL which serialises as file://C:/... instead
  of file:///C:/...)
- v0.3 also removed the sync_dir platform split — syncDir on info.go now
  fails on Windows because FlushFileBuffers requires a writable handle

Additional Windows-only bugs discovered during validation:
- CREATE_NEW_PROCESS_GROUP was missing from detached_windows.go
  spawnDetachedLoggedProcess. Unix uses Setpgid: true (detached_unix.go:58)
  but the Windows sibling forgot the equivalent, so the daemon inherits
  the parent CLI's console group and dies on parent exit.
- procutil.StartedAt crashed with ACCESS_VIOLATION (0xc0000005) inside
  GetProcessTimes when passed nil pointers for exit/kernel/user times on
  some Windows builds. Pass valid Filetime pointers for all four outputs.

Files:
- internal/daemon/lock_pid_other.go (NEW): identity lockPIDPath for Unix
- internal/daemon/lock_pid_windows.go (NEW): .pid sidecar for Windows
- internal/daemon/sync_dir_unix.go (NEW): syncDir extracted from info.go
- internal/daemon/sync_dir_windows.go (NEW): syncDir no-op for Windows
- internal/daemon/lock.go: use lockPIDPath() at 3 call sites
- internal/daemon/info.go: remove syncDir (moved to platform files)
- internal/procutil/detached_windows.go: set CREATE_NEW_PROCESS_GROUP
- internal/procutil/process_started_at_windows.go: pass valid Filetime
  pointers to GetProcessTimes, guard against invalid handles
- internal/store/sqlite_config.go: add sqliteFilePath helper that prefixes
  drive-letter absolute paths with '/' so the URI serialises correctly
- internal/store/sessiondb/read_only.go: same drive-letter normalisation
  inline (package boundary)

Zero behaviour change on Unix / Linux / macOS — every fix is behind a
//go:build windows tag or an identity function on non-Windows.

Validated end-to-end on Windows 11 (build 26200) with go 1.26.5:
- compozy daemon start succeeds
- compozy daemon status reports "running" with PID + HTTP + UDS
- compozy doctor -o json returns diagnostics via the daemon socket
